DRY fetcher handler & add link prefetch

+++ client/app/routes/erp+/_admin+/employees+/$employeeId.tsx
@@ -3,12 +3,16 @@ import { LoaderFunctionArgs, ActionFunctionArgs } from '@remix-run/node';
 import { isAuthenticated } from '~/services/auth.server';
 import { data, useLoaderData, useNavigate } from '@remix-run/react';
 import { getLast7DaysStats } from '~/services/attendance.server';
-import { parseAuthCookie } from '~/services/cookie.server';
+import {
+  getUnauthorizedActionResponse,
+  parseAuthCookie,
+} from '~/services/cookie.server';
 import ContentHeader from '~/components/ContentHeader';
 import EmployeeDetail from './_components/EmployeeDetail';
 import EmployeeAttendanceList from './_components/EmployeeAttendanceList';
 import { Edit, Pen } from 'lucide-react';
 import { canAccessEmployeeManagement } from '~/utils/permission';
+import { IActionFunctionReturn } from '~/interfaces/app.interface';
 
 export const loader = async ({ request, params }: LoaderFunctionArgs) => {
   const session = await parseAuthCookie(request);
@@ -88,13 +92,19 @@ export default function EmployeeDetails() {
   );
 }
 
-export const action = async ({ request, params }: ActionFunctionArgs) => {
+export const action = async ({
+  request,
+  params,
+}: ActionFunctionArgs): IActionFunctionReturn => {
+  const { session, headers } = await isAuthenticated(request);
+  if (!session) return getUnauthorizedActionResponse(data, headers);
+
   switch (request.method) {
     case 'DELETE':
-      const { session, headers } = await isAuthenticated(request);
       await deleteEmployee(params.employeeId!, session!);
       return data(
         {
+          success: true,
           toast: {
             type: 'success' as const,
             message: 'Xóa nhân viên thành công',
@@ -105,15 +115,13 @@ export const action = async ({ request, params }: ActionFunctionArgs) => {
     default:
       return data(
         {
+          success: false,
           toast: {
             type: 'error' as const,
             message: 'Phương thức không hợp lệ',
           },
         },
-        {
-          status: 405,
-          statusText: 'Method Not Allowed',
-        },
+        { headers, status: 405, statusText: 'Method Not Allowed' },
       );
   }
 };
